+++ src/routes/forgot.svelte
@@ -2,6 +2,7 @@
   import { onMount } from 'svelte'
   import { goto } from '$app/navigation'
   import { toast } from '../stores'
+  import { focusOnFirstError } from '$lib/focus'
 
   export const prerender = true
 
@@ -14,6 +15,8 @@
   })
 
   const sendPasswordReset = async () => {
+    const form = document.forms['forgot']
+
     message = ''
     if (email.toLowerCase().includes('gmail.com')) {
       return message = 'GMail passwords must be reset on Manage Your Google Account.'
@@ -45,12 +48,13 @@
 <div class="d-flex justify-content-center mt-5">
   <div class="card login">
     <div class="card-body">
-      <form autocomplete="on" novalidate>
+      <form id="forgot" autocomplete="on" novalidate>
         <h4><strong>Forgot password</strong></h4>
         <p>Hey, you're human. We get it.</p>
         <div class="mb-3">
           <label class="form-label" for="email">Email</label>
-          <input bind:this={focusedField} bind:value={email} type="email" id="email" class="form-control is-large" placeholder="Email" autocomplete="email"/>
+          <input bind:this={focusedField} bind:value={email} type="email" id="email" class="form-control is-large" required placeholder="Email" autocomplete="email"/>
+          <div class="invalid-feedback">Email address required</div>
         </div>
         {#if message}
           <p class="text-danger">{message}</p>