+++ src/routes/register.svelte
@@ -17,6 +17,7 @@
   import { goto } from '$app/navigation'
   import { page, session } from '$app/stores'
   import useAuth from '$lib/auth'
+  import { focusOnFirstError } from '$lib/focus'
 
   const { initializeSignInWithGoogle, registerLocal } = useAuth(page, session, goto)
 
@@ -29,9 +30,12 @@
     email: '',
     phone: ''
   }
-  let displayedError
+  let confirmPassword: HTMLInputElement
+  let displayedError: string
 
   async function register() {
+    const form = document.forms['register']
+
     displayedError = undefined
     try {
       await registerLocal(user)
@@ -49,6 +53,12 @@
       { theme: 'filled_blue', size: 'large', width: '367' }  // customization attributes
     )
   })
+
+
+  const passwordMatch = () => {
+    if (!user.password) user.password = ''
+    return user.password == confirmPassword.value
+  }
 </script>
 
 <svelte:head>
@@ -60,26 +70,35 @@
     <div class="card-body">
       <h4><strong>Register</strong></h4>
       <p>Welcome to our community.</p>
-      <form autocomplete="on" novalidate class="mt-3">
+      <form id="register" autocomplete="on" novalidate class="mt-3">
         <div class="mb-3">
           <div id="googleButton"></div>
         </div>
         <div class="mb-3">
           <label class="form-label" for="email">Email</label>
-          <input bind:this={focusedField} type="email" class="form-control is-large" bind:value={user.email} placeholder="Email" id="email" autocomplete="email"/>
+          <input bind:this={focusedField} type="email" class="form-control is-large" bind:value={user.email} required placeholder="Email" id="email" autocomplete="email"/>
+          <div class="invalid-feedback">Email address required</div>
         </div>
         <div class="mb-3">
           <label class="form-label" for="password">Password</label>
-          <input type="password" id="password" class="form-control is-large" bind:value={user.password} placeholder="Password"/>
+          <input type="password" id="password" class="form-control is-large" bind:value={user.password} required minlength="8" maxlength="80" placeholder="Password" autocomplete="new-password"/>
+          <div class="invalid-feedback">Password with 8 chars or more required</div>
+          <div class="form-text">Password minimum length 8, must have one capital letter, 1 number, and one unique character.</div>
+        </div>
+        <div class="mb-3">
+          <label class="form-label" for="password">Confirm password</label>
+          <input type="password" id="password" class="form-control is-large" bind:this={confirmPassword} required minlength="8" maxlength="80" placeholder="Password (again)" autocomplete="new-password"/>
           <div class="form-text">Password minimum length 8, must have one capital letter, 1 number, and one unique character.</div>
         </div>
         <div class="mb-3">
           <label class="form-label" for="firstName">First name</label>
-          <input bind:value={user.firstName} class="form-control" id="firstName" placeholder="First name" autocomplete="given-name"/>
+          <input bind:value={user.firstName} class="form-control" id="firstName" placeholder="First name" required autocomplete="given-name"/>
+          <div class="invalid-feedback">First name required</div>
         </div>
         <div class="mb-3">
           <label class="form-label" for="lastName">Last name</label>
-          <input bind:value={user.lastName} class="form-control" id="lastName" placeholder="Last name" autocomplete="family-name"/>
+          <input bind:value={user.lastName} class="form-control" id="lastName" placeholder="Last name" required autocomplete="family-name"/>
+          <div class="invalid-feedback">Last name required</div>
         </div>
         <div class="mb-3">
           <label class="form-label" for="phone">Phone</label>