+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 # Backlog
 
-* [Medium] Add validation on /src/routes/auth/reset/[token].svelte to make sure passwords match.
+* [Medium] Add validation to the following forms (use focusOnFirstError)
+  - [token].svelte
+  - forgot.svelte
+  - register.svelte
+  - login.svelte
+  - profile.svelte
+* [Low] Add Google reCaptcha 3